requirements.txt
firecrawl-py>=1.5.0
python-dotenv>=1.0.0
pydantic>=2.0.0
